Add support for Flexmark TOC extension (#63)
This patch adds a new dependency on flexmark-ext-toc and adds the
extension to the Flexmark parser. This enables a table of contents node
to be included in the markdown IR by including `[TOC]` in the markdown
content. Optionally a list of heading levels to be included can be
specified like so: `[TOC levels=1-3,5]`. The new IR node is named
:markdown/table-of-contents.

The patch also includes a lowering function to Hiccup. This function is
unique because it requires the full IR representation as well as the IR
node and thus requires two arguments. It analyzes the full IR for
headings and creates a nested unordered list where each heading is
linked to via their `id` values which were already present in the
heading lowering function.